style: Fix formatting and style issues  and viewItem components refactored code, created viewitem folder and made detail and review component.
